requirements.txt
torch>=1.7.0
 torchvision>=0.8.1
 numpy>=1.19.2
 pillow>=8.0.1
 scikit-image>=0.17.2
 tqdm>=4.54.1
 pytorch-wavelets>=1.3.0
 pytorch-msssim>=0.2.1
 lpips>=0.1.4
 matplotlib>=3.3.3
 PyYAML>=5.3.1
 timm>=0.4.12





colorama          0.4.6
contourpy         1.3.2
cycler            0.12.1
filelock          3.18.0
fonttools         4.58.5
fsspec            2025.5.1
imageio           2.37.0
Jinja2            3.1.6
kiwisolver        1.4.8
lazy_loader       0.4
lpips             0.1.4
MarkupSafe        3.0.2
matplotlib        3.10.3
mpmath            1.3.0
networkx          3.4.2
numpy             2.2.6
packaging         25.0
pillow            11.3.0
pip               25.1
pyparsing         3.2.3
python-dateutil   2.9.0.post0
pytorch-msssim    1.0.0
pytorch-wavelets  1.3.0
PyWavelets        1.8.0
PyYAML            6.0.2
scikit-image      0.25.2
scipy             1.15.3
setuptools        78.1.1
six               1.17.0
sympy             1.14.0
tifffile          2025.5.10
timm              0.4.12
torch             2.7.1
torchvision       0.22.1
tqdm              4.67.1
typing_extensions 4.14.1
wheel             0.45.1